fix: improve typing for multi-argument callback functions for `to`
The `to` callback function can be as simple as a two-argument function
receiving the matched string and the filename it was matched in. However
if `from` is a regular expression, the function arguments may
additionally contain captures, the offset for the matched substring, the
whole string being inspected, etc. This type definition ensures those
functions are accepted, but require providing additional type
information.