feat: accept multiple `to` replacements
The underlying `replace-in-file` package accepts multiple replacements
to be passed as an array. This functionality is made available for
multiple basic string replacements and callback functions. When using
multiple replacements the number of replacements needs to be the same as
the number of matchers. Otherwise only a subset of the replacements is
used, e.g. it's not possible to use a single string matcher and provide
multiple replacements.